+++ R/cliapp-docs.R
@@ -50,13 +50,13 @@
 #' * `file` for a file name.
 #' * `fun` for a function name.
 #' * `key` for a keyboard key.
+#' * `obj_type_friendly` formats the type of an R object in a readable way,
+#'   and it should be used with `{}`, see an example below.
 #' * `or` changes the string that separates the last two elements of
 #'   collapsed vectors (see below) from "and" to "or".
 #' * `path` for a path (essentially the same as `file`).
 #' * `pkg` for a package name.
 #' * `strong` for strong importance.
-#' * `typeof` formats the type of an R object in a readable way, and it
-#'   should be used with `{}`, see an example below.
 #' * `url` for a URL.
 #' * `var` for a variable name.
 #' * `val` for a generic "value".
@@ -73,7 +73,7 @@
 #' cli_li("An email address: {.email bugs.bunny@acme.com}.")
 #' cli_li("A URL: {.url https://acme.com}.")
 #' cli_li("An environment variable: {.envvar R_LIBS}.")
-#' cli_li("`mtcars` is {.typeof {mtcars}}")
+#' cli_li("`mtcars` is {.obj_type_friendly {mtcars}}")
 #' cli_end(ul)
 #' ```
 #'