+++ NEWS.md
@@ -1,7 +1,7 @@
 # cli (development version)
 
-* New `{.typeof}` inline style to format the type of an R object in a user
-  friendly way (#463).
+* New `{.obj_type_friendly}` inline style to format the type of an R object
+  in a user friendly way (#463).
 
 * Improved vector collapsing behavior. cli now shows both the beginning
   and end of the collapsed vector, by default (#419).